feat(tax-csv): support user can upload csv to calculate tax API
changelog
1. change float32 type to float64
2. add calculate tax from csv file api

-----------------------
HTTP Method: POST
URL: "/tax/calculations/upload-csv"
Content-Type: "multipart/form-data"
Form-Data:
    `{ "taxFile": "csv file" }`
----------------------

+++ handlers/admin.go
@@ -14,7 +14,7 @@ type AdminHandlers struct {
 }
 
 type AdminServicer interface {
-	ValidateDeductionRequest(slug string, amount float32) error
+	ValidateDeductionRequest(slug string, amount float64) error
 	UpdateDeductionConfig(deduction models.DeductionRequest) (models.Deduction, error)
 }
